src/requirements.txt
mysql_connector_repackaged==0.3.1
openpyxl==3.0.10
pandas==1.5.3
pyodbc==4.0.34
pywin32==305.1
SQLAlchemy==1.4.39
